Added paginate on project controller and fixed blades

+++ app/Http/Controllers/Admin/ProjectController.php
@@ -20,7 +20,13 @@ class ProjectController extends Controller
      */
     public function index()
     {
-        $projects = Project::all();
+        $currentUserId = Auth::id();
+        if ($currentUserId == 1) {
+            $projects = Project::all();
+            // $projects = Project::paginate(3);
+        } else {
+            $projects = Project::where('user_id', $currentUserId)->paginate(3);
+        }
         return view('admin.projects.index', compact('projects'));
     }
 
@@ -60,14 +66,21 @@ class ProjectController extends Controller
      */
     public function show(Project $project)
     {
-        return view('admin.projects.show', compact('project'));
+        $currentUserId = Auth::id();
+        if ($currentUserId == $project->user_id || $currentUserId == 1) {
+            return view('admin.projects.show', compact('project'));
+        }
+        abort(403);
     }
-
     /**
      * Show the form for editing the specified resource.
      */
     public function edit(Project $project)
     {
+        $currentUserId = Auth::id();
+        if ($currentUserId != $project->user_id && $currentUserId != 1) {
+            abort(403);
+        }
         $categories = Category::all();
         $technologies = Technology::all();
         return view('admin.projects.edit', compact('project', 'categories', 'technologies'));
@@ -78,6 +91,10 @@ class ProjectController extends Controller
      */
     public function update(UpdateProjectRequest $request, Project $project)
     {
+        $currentUserId = Auth::id();
+        if ($currentUserId != $project->user_id && $currentUserId != 1) {
+            abort(403);
+        }
         $data = $request->validated();
         $data['slug'] = $project->slug;
 
@@ -111,6 +128,10 @@ class ProjectController extends Controller
      */
     public function destroy(Project $project)
     {
+        $currentUserId = Auth::id();
+        if ($currentUserId != $project->user_id && $currentUserId != 1) {
+            abort(403);
+        }
         $project->technologies()->sync([]);
         if ($project->image) {
             Storage::delete($project->image);
